Code refactoring

+++ src/Audio/Codec/AbstractCodec.hpp
@@ -45,7 +45,7 @@ namespace SilentMedia {
           virtual void play(const string &fileId, bool resume = false) = 0;
 
           virtual float getSeek(const string &fileId) = 0;
-          virtual void setSeek(const string &fileId, const float &seekVal) = 0;
+          virtual void setSeek(const string &fileId, const double &seekVal) = 0;
         protected:
       };
     }
